[Close #207] New ActionSupport module.
Initially contains four classes de-duplicated from
PostsController::Actions::Create and UsersController::Actions::Create
(which have been updated to use the relocated code).

RSpec: 530+25+11+9+9+18+5+2+5+11+5+5+9+39+24+30 (737) examples, 0
failures
RuboCop: 242 files inspected, no offenses found
