+++ app/controllers/users_controller/action/create/internals/password_matcher.rb
@@ -1,5 +1,5 @@
 
-require_relative 'data_object_failure'
+require 'action_support/data_object_failure'
 
 # UsersController: actions related to Users within our "fancy" blog.
 class UsersController < ApplicationController
@@ -12,6 +12,8 @@ class UsersController < ApplicationController
       module Internals
         # Match password/confirmation; raise error on mismatch.
         class PasswordMatcher
+          include ActionSupport
+
           def initialize(user_data)
             @user_data = user_data
             @password_mismatch_message = 'Password must match the password' \
